fix: move `_IterXXX`'s definition
