+++ monad_std/iter/impl/__init__.py
@@ -10,7 +10,9 @@ from .fuse import Fuse
 from .inspect import Inspect
 from .intersperse import Intersperse, IntersperseWith
 from .map import Map, MapWhile, MapWindows
+from .once import OnceWith
 from .peekable import Peekable
+from .repeat import Repeat
 from .scan import Scan
 from .skip import Skip
 from .take import Take, TakeWhile
